refactor : 사용하지 않는 메소드 제거
