+++ src/main/java/controller/QuitExecutor.java
@@ -4,6 +4,6 @@ public class QuitExecutor implements MenuExecutor {
 
     @Override
     public void execute() {
-
+        //종료 로직
     }
 }
