+++ src/main/java/domain/AttendanceRecords.java
@@ -41,14 +41,6 @@ public record AttendanceRecords(
                 .orElseThrow(() -> new IllegalArgumentException("[ERROR] 입력한 날짜의 출석 기록이 없습니다."));
     }
 
-    public void updateAttendanceByDate(LocalDate attendanceDate, int hour, int minute, AttendanceStatus updateAttendanceStatus) {
-        Attendance originAttendance = attendances.stream()
-                .filter(attendance -> attendance.isEqualDate(attendanceDate))
-                .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("[ERROR] 입력한 날짜의 출석 기록이 없습니다."));
-        originAttendance.updateAttendance(hour, minute, updateAttendanceStatus);
-    }
-
     public static AttendanceRecords create(List<Attendance> attendances) {
         return new AttendanceRecords(attendances);
     }
