+++ src/test/java/AttendanceRecordsTest.java
@@ -60,64 +60,6 @@ public class AttendanceRecordsTest {
         assertThat(foundAttendance).isSameAs(attendance);
     }
 
-    @DisplayName("조회하고자 하는 날의 출석기록이 없다면 예외를 발생시킨다")
-    @Test
-    void retrieveException() {
-        //given
-        LocalDate localDate = LocalDate.of(2024, 12, 13);
-        Attendance attendance = AttendanceFixture.createAttendance(localDate, 10, 0);
-        AttendanceRecords attendanceRecords = AttendanceRecords.create(new ArrayList<>(List.of(attendance)));
-
-        LocalDate anotherDate = LocalDate.of(2024, 12, 12);
-        //when && then
-        assertThatThrownBy(() -> attendanceRecords.updateAttendanceByDate(anotherDate, 0, 0, null))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 입력한 날짜의 출석 기록이 없습니다.");
-    }
-
-    @DisplayName("원하는 날짜의 출석을 수정할 수 있다")
-    @Test
-    void updateByDate() {
-        //given
-        LocalDate localDate = LocalDate.of(2024, 12, 13);
-        Attendance attendance = AttendanceFixture.createAttendance(localDate, 10, 0);
-        AttendanceRecords attendanceRecords = AttendanceRecords.create(new ArrayList<>(List.of(attendance)));
-
-        int hour = 10;
-        int minute = 0;
-        AttendanceStatus updateAttendanceStatus = AttendanceStatus.ATTENDANCE;
-
-        //when
-        attendanceRecords.updateAttendanceByDate(localDate, hour, minute, updateAttendanceStatus);
-
-        //then
-        SoftAssertions.assertSoftly(softly -> {
-            softly.assertThat(attendance).extracting("attendanceTime").extracting("hour").isEqualTo(10);
-            softly.assertThat(attendance).extracting("attendanceTime").extracting("minute").isEqualTo(0);
-            softly.assertThat(attendance).extracting("attendanceStatus").isEqualTo(AttendanceStatus.ATTENDANCE);
-        });
-    }
-
-    @DisplayName("수정하고자 하는 날의 출석기록이 없다면 예외를 발생시킨다")
-    @Test
-    void updateException() {
-        //given
-        LocalDate localDate = LocalDate.of(2024, 12, 13);
-        Attendance attendance = AttendanceFixture.createAttendance(localDate, 10, 0);
-        AttendanceRecords attendanceRecords = AttendanceRecords.create(new ArrayList<>(List.of(attendance)));
-
-        int hour = 10;
-        int minute = 0;
-        AttendanceStatus updateAttendanceStatus = AttendanceStatus.ATTENDANCE;
-
-        LocalDate anotherDate = LocalDate.of(2024, 12, 12);
-        //when && then
-        assertThatThrownBy(
-                () -> attendanceRecords.updateAttendanceByDate(anotherDate, hour, minute, updateAttendanceStatus))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 입력한 날짜의 출석 기록이 없습니다.");
-    }
-
     @DisplayName("입력 날짜의 전날까지 출석 기록 조회한다")
     @Test
     void retrieveAllAttendanceUntilDate() {
